refactor: move and rename method
